[bug-fix](client): fix firebase issue where multiple instances were initializing
